# nixos-modules: add closure size + startup time optimization