--- a/nixos-modules/microvm/default.nix
+++ b/nixos-modules/microvm/default.nix
@@ -16,6 +16,7 @@ in
     ./system.nix
     ./mounts.nix
     ./graphics.nix
+    ./optimization.nix
   ];
 
   config = {