@lf- 2023-10-01
oh, the nonsense device times are actually just, the entire initrd basically. https://github.com/systemd/systemd/issues/29010